+++ app/src/main/java/com/app/arcabyolimpo/domain/repository/workshops/WorkshopRepository.kt
@@ -11,4 +11,7 @@ interface WorkshopRepository {
 
     suspend fun addWorkshop(newWorkshop: WorkshopDto): Workshop
 
+
+    companion object
+
 }